Implement fetch data from DB with thunk
